# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FX Scalper — ICT × Jev × Deriv",
  "summary": "Transcripts → Jev-tagged ICT concepts → typed strategy graph → Deriv backtests and demo execution.",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/trading/engine/start": {
   "post": {
    "tags": [
     "trading"
    ],
    "summary": "Start the engine (paper = simulated fills; demo = Deriv demo orders)",
    "operationId": "engine_start_api_trading_engine_start_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EngineConfig"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Engine Start Api Trading Engine Start Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/trading/calibrate": {
   "post": {
    "tags": [
     "trading"
    ],
    "summary": "Measure Deriv's real commission, accepted multipliers and Rise/Fall payout",
    "operationId": "calibrate_api_trading_calibrate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CalibrateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Calibrate Api Trading Calibrate Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/trading/signals": {
   "get": {
    "tags": [
     "trading"
    ],
    "summary": "Evaluated setups (accepted and rejected) with node-by-node reasons",
    "operationId": "signals_api_trading_signals_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Signals Api Trading Signals Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/trading/trades": {
   "get": {
    "tags": [
     "trading"
    ],
    "summary": "Trade journal (paper/demo/real), newest first",
    "operationId": "trades_api_trading_trades_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "mode",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mode"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Trades Api Trading Trades Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/trading/trades.csv": {
   "get": {
    "tags": [
     "trading"
    ],
    "summary": "Trade journal as CSV",
    "operationId": "trades_csv_api_trading_trades_csv_get",
    "parameters": [
     {
      "name": "mode",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mode"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/trading/signals/{signal_id}/confidence": {
   "get": {
    "tags": [
     "trading"
    ],
    "summary": "The ensemble confidence JSON for one signal",
    "operationId": "signal_confidence_api_trading_signals__signal_id__confidence_get",
    "parameters": [
     {
      "name": "signal_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Signal Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Signal Confidence Api Trading Signals  Signal Id  Confidence Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/trading/signals/{signal_id}/pine": {
   "get": {
    "tags": [
     "trading"
    ],
    "summary": "Pine Script v6 strategy for a signal's entry, stop, target and risk",
    "operationId": "signal_pine_api_trading_signals__signal_id__pine_get",
    "parameters": [
     {
      "name": "signal_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Signal Id"
      }
     },
     {
      "name": "risk_amount",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Risk Amount"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "text/plain": {
        "schema": {
         "type": "string"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/trading/pine": {
   "post": {
    "tags": [
     "trading"
    ],
    "summary": "Pine Script v6 strategy for any plan",
    "operationId": "pine_from_plan_api_trading_pine_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PinePlan"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "text/plain": {
        "schema": {
         "type": "string"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CalibrateRequest": {
    "properties": {
     "symbols": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Symbols"
     },
     "rise_fall_minutes": {
      "type": "integer",
      "title": "Rise Fall Minutes",
      "default": 15
     }
    },
    "type": "object",
    "title": "CalibrateRequest"
   },
   "EngineConfig": {
    "properties": {
     "strategy_id": {
      "type": "string",
      "title": "Strategy Id",
      "default": "ict_2022_model"
     },
     "version": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Version"
     },
     "symbols": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Symbols"
     },
     "mode": {
      "type": "string",
      "enum": [
       "paper",
       "demo"
      ],
      "title": "Mode",
      "default": "paper"
     },
     "evaluation": {
      "type": "string",
      "enum": [
       "code",
       "jev",
       "laya",
       "ensemble"
      ],
      "title": "Evaluation",
      "default": "code"
     },
     "risk_amount": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Risk Amount",
      "description": "Account currency lost at the planned stop (1R)",
      "default": 1.0
     },
     "currency": {
      "type": "string",
      "title": "Currency",
      "default": "USD"
     }
    },
    "type": "object",
    "title": "EngineConfig"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PinePlan": {
    "properties": {
     "symbol": {
      "type": "string",
      "title": "Symbol"
     },
     "direction": {
      "type": "string",
      "pattern": "^(long|short)$",
      "title": "Direction"
     },
     "entry": {
      "type": "number",
      "title": "Entry"
     },
     "stop": {
      "type": "number",
      "title": "Stop"
     },
     "target": {
      "type": "number",
      "title": "Target"
     },
     "risk_amount": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Risk Amount",
      "default": 1.0
     },
     "signal_time": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Signal Time"
     },
     "expires_at": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "symbol",
     "direction",
     "entry",
     "stop",
     "target"
    ],
    "title": "PinePlan"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 },
...
}